Fix crash while doing emergency call in airplane mode.
When emergency call is attempted in APM, radio is powered on first.
If there is voice radio tech switch after radio on, the old
phone instance is disposed and a new one is created.
EmergencyCallhelper crashes when trying to use the disposed phone
instance.Use PhoneProxy object instead of PhoneBase which will take
care of radio tech change.

Change-Id: Ie7f367f5b9e29c6870cabc8c3cb832ebc49f8c3a
CRs-Fixed: 534652